+++ tools/generator/model/src/main/java/is/codion/tools/generator/model/package-info.java
@@ -19,8 +19,8 @@
 /**
  * Package configuration values:
  * <ul>
- * <li>{@link is.codion.tools.generator.model.DomainGeneratorModel#DEFAULT_DOMAIN_PACKAGE}
- * <li>{@link is.codion.tools.generator.model.DomainGeneratorModel#DEFAULT_SOURCE_DIRECTORY}
+ * <li>{@link is.codion.tools.generator.model.DomainGeneratorModel#DOMAIN_PACKAGE}
+ * <li>{@link is.codion.tools.generator.model.DomainGeneratorModel#SOURCE_DIRECTORY}
  * </ul>
  */
 package is.codion.tools.generator.model;